Java based application on medical pharmacy.
Medical Pharmacy management project is the process of managing the
medicine stock and selecting the suitable medicine to the illness that
mention by patient or customer. The core of pharmacist profession is the
maintenance of quality and the subsequent implication for patient care.

The domain of this project is in Healthcare. Within the growth of the
Medical Technology, the system developer takes this opportunity to help
the
pharmacist for managing their stock and select the medicine using
computer
program.

This system also does the selling, ordering  and buying the medicine.
Selling process is involved because to relate the alert warning message
to the stock condition. Additional, Medical Pharmacy Project  provide
the
information of the medicine that is sold and ordered.

The modules involve are medicine management module, medicine selector
module, selling process module, medicine list module, customer
information module, order module, stock of medicine sales module.